🧹 Reduce PositionRow props by utilizing zustand store directly for simple actions
